Split public interface for token providers into separate files.
