+++ Chatkit/Authentication/Types/TokenProvider.swift
@@ -1,0 +1,18 @@
+import Foundation
+import PusherPlatform
+
+// FIXME: Temporary
+extension String: Error { }
+
+public typealias TokenProvider = PPTokenProvider
+public typealias TokenProviderResult = PPTokenProviderResult
+
+public typealias AsyncKeyValueCall = (CredentialsCompletionHandler) -> Void
+public typealias CredentialsCompletionHandler = (KeyValueResult) -> Void
+
+public enum KeyValueResult {
+    
+    case success([String: String])
+    case error(Error)
+    
+}
